firmware/testing/python/requirements.txt
astroid==2.9.3
guizero==1.3.0
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.6.1
pick==1.2.0
Pillow==9.0.1
platformdirs==2.5.1
pylint==2.12.2
pyserial==3.5
toml==0.10.2
typing-extensions==4.1.1
wrapt==1.13.3
